Add no_proxy_uris param to apache::vhost::proxy
Add a mechanism to prevent certain URI's from being proxied.

This is particularly useful if you want to specify a proxy vhost pointed
at a webapp that installs itself under /SOME_URI in Apache's global
configuration. This is common on Debian and derivatives. Adding a vhost
for these is now as simple as:

   apache::vhost::proxy { 'manage.example.com':
     port          => 80,
     dest          => 'http://localhost/openstack-dashboard',
     no_proxy_uris => '/openstack-dashboard'
   }

Without this, you can easily get caught in an infinite, internal
redirect loop.